ktp-scanner/requirements.txt
# KTP Scanner - DL/ML microservice
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
Pillow>=10.0.0
numpy>=1.24.0

# OCR: EasyOCR (Indonesian + English, works well on Windows/Linux)
easyocr>=1.7.0
# Optional: PaddleOCR for higher accuracy (uncomment if preferred)
# paddlepaddle>=2.5.0
# paddleocr>=2.7.0

# Validation script
requests>=2.31.0
